Process review comments
- Update docstrings where applicable

  * https://github.com/scality/ScalitySproxydSwift/pull/86#discussion_r31435437
  * https://github.com/scality/ScalitySproxydSwift/pull/86#discussion_r31435462

- Move `ClientCollection` and `NoClientAvailable` into `http_utils`

  * https://github.com/scality/ScalitySproxydSwift/pull/86#discussion_r31435897

- Add code comments to `_get_client_for_policy`

  * https://github.com/scality/ScalitySproxydSwift/pull/86#discussion_r31437047
  * https://github.com/scality/ScalitySproxydSwift/pull/86#discussion_r31437125
  * https://github.com/scality/ScalitySproxydSwift/pull/86#discussion_r31437218

See: https://github.com/scality/ScalitySproxydSwift/pull/86
