+++ test/unit/test_diskfile.py
@@ -26,8 +26,8 @@ import swift.common.utils
 
 from swift_scality_backend.diskfile import DiskFileWriter, \
     DiskFileReader, DiskFile, DiskFileManager
-from swift_scality_backend.diskfile import ClientCollection
 from scality_sproxyd_client.exceptions import SproxydHTTPException
+from swift_scality_backend.http_utils import ClientCollection
 from scality_sproxyd_client.sproxyd_client import SproxydClient
 from . import utils
 from .utils import make_client_collection
